+++ lib/fluttersdk_wind.dart
@@ -1,7 +1,8 @@
-library fluttersdk_wind;
+library;
 
 export 'src/core/platform_service.dart';
 export 'src/parser/parsers/background_parser.dart';
+export 'src/parser/parsers/border_parser.dart';
 export 'src/parser/parsers/display_parser.dart';
 export 'src/parser/parsers/flexbox_grid_parser.dart';
 export 'src/parser/parsers/margin_parser.dart';